Add files via upload
The face mask detection project presents a solution to a binary image classification problem using a convolutional neural network (CNN). The goal is to classify images of people with and without masks. The model is trained on a dataset that contains images of people with and without masks. The performance of the model is evaluated using various metrics, such as confusion matrix, classification report, and AUC curve.